minor improvements and setup fix
