+++ src/visualize.py
@@ -2,7 +2,8 @@ import json
 import pandas as pd
 import matplotlib.pyplot as plt
 import numpy as np
-import algorithms as algorithms
+import os
+import algorithms_ai as algorithms_ai
 
 # read settings.json
 
@@ -10,14 +11,17 @@ with open("settings.json") as f:
     parameters = json.load(f)
     subsets = parameters['subsets']
     resolution = parameters['resolution']
+    data_dir = parameters['data_dir']
+    path_points_name = parameters['path_points_name']
+    path_name = parameters['path_name']
 
 res = subsets / resolution
 # read files
-path_points = pd.read_csv(algorithms.path_points_path)
-path = pd.read_csv(algorithms.path_path)
+path_points = pd.read_csv(os.path.join(data_dir, f"{path_points_name}.csv"))
+path = pd.read_csv(os.path.join(data_dir, f"{path_name}.csv"))
 
 # plot settings
-boldness = 0.01
+boldness = 0.05
 
 # plot the path
 fig = plt.figure()
